Change file image depending on file type
