+++ libs/media/src/lib/components/upload/widget/upload-widget.component.html
@@ -14,7 +14,7 @@
         <ng-container *ngFor="let task$ of tasks.$ | async; let i = index">
           <ng-container [ngSwitch]="task$ | state | async">
             <div fxLayout="row" fxLayoutGap="8px">
-              <img [src]="getFileType(task$.task.snapshot.ref.name)">
+              <img [asset]="task$.task.snapshot.ref.name | fileTypeImage">
 
               <div fxLayout="column" fxLayoutAlign="start start">
                 <caption class="caption">{{task$.task.snapshot.ref.name | maxLength: 25}}</caption>
