+++ libs/media/src/lib/components/upload/widget/upload-widget.component.ts
@@ -20,26 +20,6 @@ export class UploadWidgetComponent {
     @Inject('tasks') public tasks: BehaviorStore<AngularFireUploadTask[]>,
   ) {}
 
-  getFileType(file: string) {
-    const type = file.split('.').pop();
-    switch (type) {
-      case 'docx':
-      case 'doc':
-        return '/assets/images/dark/docx.webp';
-      case 'xls':
-      case 'xlsx':
-        return '/assets/images/dark/xls.webp';
-      case 'png':
-      case 'webp':
-      case 'jpg':
-        return '/assets/images/dark/image.webp';
-      case 'pdf':
-        return '/assets/images/dark/pdf.webp';
-      default:
-        return '/assets/images/dark/image.webp';
-    }
-  }
-
   cancel(task: AngularFireUploadTask) {
     task.resume();
     task.cancel();
