+++ libs/media/src/lib/components/upload/widget/upload-widget.module.ts
@@ -10,6 +10,8 @@ import { UploadWidgetComponent } from './upload-widget.component';
 import { MaxLengthModule } from '@blockframes/utils/pipes/max-length.pipe';
 import { TaskProgressModule } from '../../../pipes/task-progress.pipe';
 import { TaskSnapshotModule } from '../../../pipes/task-state.pipe';
+import { ImageReferenceModule } from '@blockframes/media/directives/image-reference/image-reference.module';
+import { FileTypeImagePipeModule } from '@blockframes/media/pipes/file-type-image.pipe';
 
 // Material
 import { MatExpansionModule } from '@angular/material/expansion';
@@ -27,6 +29,10 @@ import { MatDividerModule } from '@angular/material/divider';
     MaxLengthModule,
     TaskProgressModule,
     TaskSnapshotModule,
+    FileTypeImagePipeModule,
+
+    // Blockframes
+    ImageReferenceModule,
 
     // Material
     MatExpansionModule,
